Update
Update RenderMeshOnScreen

+++ StudioProject2/Application/Source/SceneMuseum.cpp
@@ -1035,7 +1035,7 @@ void SceneMuseum::RenderTextOnScreen(Mesh* mesh, std::string text, Color color, 
 
 }
 
-void SceneMuseum::RenderMeshOnScreen(Mesh* mesh, int x, int y, int sizex, int sizey)
+void SceneMuseum::RenderMeshOnScreen(Mesh* mesh, float x, float y, float sizex, float sizey)
 {
 	glDisable(GL_DEPTH_TEST);
 	Mtx44 ortho;
